docs: transfer the oddDozColum constant to a separate module
