+++ roulette/models/consts/oddDozColum.js
@@ -1,0 +1,15 @@
+// шансы, дюжины и колонки:
+export const oddDozColum = {
+    dozen_1: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    dozen_2: [13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24],
+    dozen_3: [25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36],
+    column_1: [1, 4, 7, 10, 13, 16, 19, 22, 25, 28, 31, 34],
+    column_2: [2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 32, 35],
+    column_3: [3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36],
+    small: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18],
+    big: [19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36],
+    even: [2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32, 34, 36],
+    odd: [1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 33, 35],
+    red: [1, 3, 5, 7, 9, 12, 14, 16, 18, 19, 21, 23, 25, 27, 30, 32, 34, 36],
+    black: [2, 4, 6, 8, 10, 11, 13, 15, 17, 20, 22, 24, 26, 28, 29, 31, 33, 35]
+};
