+++ roulette/models/fieldAndTrackModel.js
@@ -4,7 +4,7 @@
 // import { payoutRatios } from "./consts/payoutRatios.js";
 // import { chipsNeededForABet } from "./menuModel.js";
 // import { countPositions } from "./menuModel.js";
-// import { oddDozColum } from "./menuModel.js";
+// import { oddDozColum } from "./consts/oddDozColum.js";
 // import { rouletteSeries } from './consts/rouletteSeries.js';
 
 import { rouletteNumber } from "./consts/rouletteNumber.js";
